refactor: RepositorySubmissionBox 리팩토링
