+++ apps/client/types/entities/myAssignment.ts
@@ -1,4 +1,4 @@
 export type RepositorySubmissionStatusType =
   | "EDITING"
   | "SUBMITTED"
-  | "WARNING_WITH_NO_URL";
+  | "EDITING_WITH_WARNING";
